fix bug where article with fae_image presence validation can be saved without image
This bug happens when a validation error causes the form's update to fail.
The image that was initially uploaded stays "cached" in the form,
and saving the form a second time, tricks the backend into thinking
the image is present when it is not.

also, even though there is a presence validation on the image, there is no
feedback to the user. So I added there error message above the image upload
button.